requirements.txt
python-docx
openai
python-dotenv
colorama
